Remove binary files
The `main.pdf` file has been removed as it changes every time it is
compiled whether or not the source TeX has changed. Including it is
considered bad practice.

The `flatex` linux binary has been removed and the `flatex.c` source
code has instead been included (as it should have been from the
begining). The Makefile has been updated to build `flatex` if it is
missing. The scripts README has also been updated to reflect this
change.